Sort array containing all literals by number of clauses the literal occurs in
This way when we pick a random literal to assign a value to we pick the literal
with the most occurrences that is unassigned.
In general it is smarter to assign values to literals with more occurrences
because it will affect more clauses, which will hopefully lead to finding
a solution faster.